+++ src/main/java/io/split/android/client/service/sseclient/notifications/MySegmentsV2PayloadDecoder.java
@@ -46,7 +46,7 @@ public class MySegmentsV2PayloadDecoder {
         return (keyMap[internal] & 1 << offset) != 0;
     }
 
-    public static BigInteger hashKey(String key) {
+    public BigInteger hashKey(String key) {
         return MurmurHash3.unsignedHash128x64(key.getBytes(StringHelper.defaultCharset()))[0];
     }
